iow-uprn-lat-log-coefficient-pipeline/adapter/requirements.txt
python-dotenv
telicent-lib==1.0.1